Refactor the examples by categorizing them into broader themes.
