+++ docs/src/examples/cutting_stock_column_generation.jl
@@ -3,7 +3,7 @@
 # v.2.0. If a copy of the MPL was not distributed with this file, You can       #src
 # obtain one at https://mozilla.org/MPL/2.0/.                                   #src
 
-# # Cutting stock
+# # Cutting stock column generation
 
 # This example solves the cutting stock problem (sometimes also called the
 # cutting rod problem) using a column-generation technique. It is based on
