--- docs/src/examples/rosenbrock.jl
@@ -1,30 +1,0 @@
-# Copyright 2017, Iain Dunning, Joey Huchette, Miles Lubin, and contributors    #src
-# This Source Code Form is subject to the terms of the Mozilla Public License   #src
-# v.2.0. If a copy of the MPL was not distributed with this file, You can       #src
-# obtain one at https://mozilla.org/MPL/2.0/.                                   #src
-
-# # The Rosenbrock function
-
-# A nonlinear example of the classical Rosenbrock function.
-
-using JuMP
-import Ipopt
-import Test
-
-function example_rosenbrock()
-    model = Model(Ipopt.Optimizer)
-    set_silent(model)
-    @variable(model, x)
-    @variable(model, y)
-    @NLobjective(model, Min, (1 - x)^2 + 100 * (y - x^2)^2)
-    optimize!(model)
-
-    Test.@test termination_status(model) == MOI.LOCALLY_SOLVED
-    Test.@test primal_status(model) == MOI.FEASIBLE_POINT
-    Test.@test objective_value(model) ≈ 0.0 atol = 1e-10
-    Test.@test value(x) ≈ 1.0
-    Test.@test value(y) ≈ 1.0
-    return
-end
-
-example_rosenbrock()
